wwc2019/email_function/requirements.txt
google-cloud-firestore>=0.4.7
sendgrid>=6.1.0
google-auth>=1.7.1